Applying S.R. 8 and 10 in attr and fix MATCH behaviour.
Match inside of the command function so that we have access to
values in lexical entry. For example, if we've kept matching inside
syntatic_command, when entering syntatic_attr we wouldn't have access to
the identifier in the symbol table anymore, so we would need to pass it
as a parameter and create a blob in syntatic_command to get some values
we don't even know we would need to use.