Adding a last_entry to syntatic_ctx.
There are many places where we currently need to store a value because
it will be overwritten after MATCH_OR_ERROR that allows use to use that
value.

For example, only a TOKEN_IDENTIFIER has a valid pointer to an entry in
the symbol table. That means we need to get the pointer, but wait after
MATCH_OR_ERROR to use it... if we access it before confirming it's an
IDENTIFIER, we will just segfault.

By copying the last_entry, we can always access it directly.