IMPORTANT: Strings always reserve 256 bytes of memory.
Strings can have at most 255 ASCII characters and a NULL terminator.

Unnitialized strings always reserve 256 bytes, because we only know what
will be inserted into them at runtime.

Previously, we only reserved the needed number of bytes per initialized
string, but that becomes a problem for two reasons:

1 - Consider the following snippet of code:
    string a := "A";
    a := "A bigger String";
This would cause an overflow. Since we initially only reserved 2 bytes
for A, but now we are trying to place a bigger number of bytes in it.

2 - Easier management of alignment can be done. We can't align to non
power of two numbers.